Extracting signatures
This adds two scripts. The first, 's2n_model.R', contains the modeling function that extracts and tests the signature (using LOOCV). The second, '04_extracting_signatures.Rmd', uses this helper function to extract the Ven and Dec-Ven signature from the 4-patient data (exp25) using various numbers of markers (3-15). Correlation of the signal with Ven and Aza+Ven is computed using the BeatAML 210 cohort. These states are all plotted versus the number of markers to decide on optimal numbers.