Create Exp28_quick_analysis.Rmd
This is the prelim analysis script for the exp28 data (hma+ven samples). In this script we compute differential expression and plot heatmaps of the most significant hits, compute and plot GSEA, and apply various signatures (from exp25 and the sorted cells) to the hma+ven samples to check if they distinguish between the outcome cohorts.

All data loaded from synapse.